requirements.txt
google-cloud-scheduler
Django
uuid